Refactor test file structure and improve import paths for consistency
