+++ src/Calendar/utils/tests/getEventsWithEventGroups.test.js
@@ -1,12 +1,13 @@
-import getEventsWithEventGroups from './getEventsWithEventGroups';
-import getMungedEvents from './getMungedEvents';
-import { MOCKED_EVENTS } from '../mocks';
+import getEventsWithEventGroups from '../getEventsWithEventGroups';
+import getMungedEvents from '../getMungedEvents';
+import { MOCKED_EVENTS } from '../../mocks';
 
 describe('Munged Event Object', () => {
   it('Must have an event from two calendars', () => {
     const mungedEvents = getMungedEvents({
       events: MOCKED_EVENTS,
       stepMinutes: 30,
+      withColumns: false,
     });
 
     const events = getEventsWithEventGroups({
@@ -14,11 +15,11 @@ describe('Munged Event Object', () => {
       selectedEventGroups: [5, 6],
     });
 
-    const hasEvent3 = events['2019-02-12'].some(event => {
+    const hasEvent3 = events['2019-02-12'].some((event) => {
       return event.id === 3;
     });
 
-    const hasEvent5 = events['2019-02-12'].some(event => {
+    const hasEvent5 = events['2019-02-11'].some((event) => {
       return event.id === 5;
     });
