+++ src/Calendar/utils/tests/getMungedEvents.test.js
@@ -1,5 +1,5 @@
-import getMungedEvents from './getMungedEvents';
-import { MOCKED_EVENTS } from '../mocks';
+import getMungedEvents from '../getMungedEvents';
+import { MOCKED_EVENTS } from '../../mocks';
 
 describe('Munged Event Object', () => {
   it('has keys that are calendar ids', () => {
@@ -16,6 +16,7 @@ describe('Munged Event Object', () => {
     const mungedEvents = getMungedEvents({
       events: MOCKED_EVENTS,
       stepMinutes: 30,
+      withColumns: false,
     });
 
     const calendar5 = mungedEvents[5];
@@ -30,6 +31,7 @@ describe('Munged Event Object', () => {
     const mungedEvents = getMungedEvents({
       events: MOCKED_EVENTS,
       stepMinutes: 30,
+      withColumns: false,
     });
 
     const event = mungedEvents[5]['2019-01-02'][0];
