+++ src/components/Toolbar/utils/tests/useFetchEvents.test.js
@@ -1,9 +1,9 @@
 import moment from 'moment';
-import { getNewRanges, getNewRangeOutsideRange } from './useFetchEvents';
+import { getNewRanges, getNewRangeOutsideRange } from '../useFetchEvents';
 
 const monthPadding = 1;
 
-const toString = date => {
+const toString = (date) => {
   return date.format('YYYY-MM-DD');
 };
 
@@ -136,21 +136,15 @@ describe('Get outside range for calendar fetch more', () => {
       moment(new Date('2020-02-29')),
     ];
 
-    selectedDates.forEach(selectedDate => {
+    selectedDates.forEach((selectedDate) => {
       const newRanges = getNewRangeOutsideRange({ selectedDate, monthPadding });
 
       const expected = {
         start: toString(
-          selectedDate
-            .clone()
-            .subtract(monthPadding, 'months')
-            .startOf('month')
+          selectedDate.clone().subtract(monthPadding, 'months').startOf('month')
         ),
         end: toString(
-          selectedDate
-            .clone()
-            .add(monthPadding, 'months')
-            .endOf('month')
+          selectedDate.clone().add(monthPadding, 'months').endOf('month')
         ),
       };
